+++ app/decorators/judge_detail_decorator.rb
@@ -1,10 +1,10 @@
-class ElementJudgeDetailDecorator < EntryDecorator
+class JudgeDetailDecorator < EntryDecorator
   def score_name
-    h.link_to_score(model.element.score)
+    h.link_to_score(model.detailable.score)
   end
 
   def skater_name
-    h.link_to_skater(model.element.score.skater)
+    h.link_to_skater(model.detailable.score.skater)
   end
 
   def panel_name