+++ app/decorators/panel_decorator.rb
@@ -1,5 +1,5 @@
 class PanelDecorator < EntryDecorator
-  def name
+  def panel_name
     h.link_to(model.name, h.panel_path(model.name))
   end
 end