+++ db/migrate/005_create_judge_details.rb
@@ -11,17 +11,18 @@ class CreateJudgeDetails < ActiveRecord::Migration[5.1]
       t.belongs_to :performed_segment
     end
 
-    create_table :element_judge_details do |t|
+    create_table :judge_details do |t|
       t.integer :number
       t.float :value
-      t.float :average
-      t.float :deviation
-      t.float :abs_deviation
+      #t.float :average
+      #t.float :deviation
+      #t.float :abs_deviation
 
-      t.belongs_to :element
+      t.references :detailable, polymorphic: true, index: true
       t.belongs_to :official
     end
 
+=begin
     create_table :component_judge_details do |t|
       t.integer :number
       t.float :value
@@ -31,6 +32,7 @@ class CreateJudgeDetails < ActiveRecord::Migration[5.1]
       t.belongs_to :component
       t.belongs_to :official
     end
+=end
 
     ################
     create_table :deviations do |t|