+++ lib/ajax_datatables/datatable/limit.rb
@@ -1,5 +1,5 @@
 module AjaxDatatables::Datatable::Limit
-  MIN_LENGTH = 10
+  MIN_LENGTH = 25
   MAX_LENGTH = 1_000
   def limit(number = MAX_LENGTH, offset = 0)
     @limit = (number.to_i.zero?) ? MIN_LENGTH : [number.to_i, MAX_LENGTH].min