+++ spec/ajax_feature_helper.rb
@@ -1,19 +1,25 @@
 module AjaxFeatureHelper
-  SLEEP_COUNT = 0.3
-
-  shared_examples :contains do |main_flag, sub_flag|
-    if main_flag
-      it { is_expected.to have_content(main.name) }
-    else
-      it { is_expected.not_to have_content(main.name) }
-    end
-    if sub_flag
-      it { is_expected.to have_content(sub.name) }
-    else
-      it { is_expected.not_to have_content(sub.name) }
+  SLEEP_COUNT = 0.5
+  shared_context :contains_all do |datatable|
+    context 'contains_all' do
+      it {
+        visit datatable_index_path(datatable)
+        datatable.data.all.each do |item|
+          expect(page.text).to have_content(item.name)
+        end
+      }
     end
   end
-  def ajax_actions(actions, path:)
+  ## shared functions
+  def get_datatable(page)
+    table_id = page.find(:css, '.dataTable')[:id]
+    page.find("##{table_id}")
+  end
+  def datatable_index_path(datatable)
+    send("#{datatable.default_model.to_s.pluralize.underscore}_path".to_sym)
+  end
+  
+  def ajax_actions(actions, path:, format: :html)
     visit path
     actions.each do |hash|
       key, value, input_type = hash.values_at(:key, :value, :input_type)
@@ -23,84 +29,86 @@ module AjaxFeatureHelper
         find("input##{key}").send_keys :tab
       when :select
         select value, from: key
-      when :click, :checkbox
-        find(key).click
+      when :click, :checkbox, :button, :submit
+        find_by_id(key).click
       end
-      # sleep 1
-      sleep SLEEP_COUNT
+      yield if block_given?
     end
+    case format
+    when :json
+      find_by_id('json_button').click
+    end
+
+    sleep SLEEP_COUNT
     page
   end
   ################
   module Filter
-    shared_context :filter do |filters, excludings: []|
-      filters.map { |filter| filter.children.presence || filter }.flatten
-        .reject { |filter| excludings.include?(filter.key) }.each do |filter|
-        include_context :ajax_filter, filter
+    shared_examples :filter do |filter, additional_actions: nil, value_func: nil, pros_operator: :eq, cons_operator: :not_eq|
+      it {
+        datatable = filter.filters.datatable
+        column = datatable.columns[filter.key] || next
+        value_func ||= lambda {|dt, key| dt.data.first.send(key) }
+        
+        value = value_func.call(datatable, filter.key)
+
+        ## pros, cons
+        arel = column.table_model.arel_table[column.table_field]
+        pros = datatable.data.where(arel.send(pros_operator, value))
+        cons = datatable.data.where(arel.send(cons_operator, value))
+
+        path = datatable_index_path(datatable)
+        actions =  [{key: filter.key, input_type: filter.input_type, value: value}]
+        actions.push(*additional_actions) if additional_actions
+
+        ajax_actions(actions, path: path)
+        table_text = get_datatable(page).text
+
+        
+        pros.each do |item|
+          expect(table_text).to have_content(item.name)
+        end
+        cons.each do |item|
+          expect(table_text).not_to have_content(item.name)
+        end
+        ajax_actions(actions, path: path, format: :json)
+        pros.each do |item|
+          expect(page.text).to have_content(item.name)
+        end
+        cons.each do |item|
+          expect(page.text).not_to have_content(item.name)
+        end
+      }
+    end
+    shared_context :filters do |datatable| #, excludings: []|
+      context 'filters' do
+        datatable.filters.map { |filter| filter.children.presence || filter }.flatten.each do |filter|
+          #.reject { |filter| excludings.include?(filter.key) }.each do |filter|
+          context filter.key do
+            it_behaves_like :filter, filter
+          end
+        end
+      end
+    end
+    shared_context :filter_with_operator do |filter, operator_key, operator_value|
+      context "#{operator_key} #{operator_value}" do
+        direction, pros_operator, cons_operator =
+                                  case operator_value
+                                  when '>' then [:asc, :gt, :lteq]
+                                  when '<' then [:desc, :lt, :gteq]
+                                  end
+        
+        actions = [{ key: operator_key, value: operator_value, input_type: :select }]
+        value_func = lambda {|dt, key| dt.data.order("#{dt.columns[key].source} #{direction}").first.send(key) }
+        it_behaves_like :filter, filter, additional_actions: actions, value_func: value_func, pros_operator: pros_operator, cons_operator: cons_operator
       end
     end
 
-    shared_context :filter_season do
-      context 'eq season' do
-        subject {
-          ajax_actions([{ key: :season_operator, value: '=', input_type: :select },
-                        { key: :season, value: main.season, input_type: :select }], path: index_path)
-        }
-        it_behaves_like :contains, true, false
-      end
-      context 'lt season' do
-        subject {
-          ajax_actions([{ key: :season_operator, value: '<', input_type: :select },
-                        { key: :season, value: main.season, input_type: :select }], path: index_path)
-        }
-        it_behaves_like :contains, false, false
-      end
-      context 'gt season' do
-        subject {
-          ajax_actions([{ key: :season_operator, value: '>', input_type: :select },
-                        { key: :season, value: main.season, input_type: :select }], path: index_path)
-        }
-        it_behaves_like :contains, false, true
-      end
-      context 'lteq season' do
-        subject {
-          ajax_actions([{ key: :season_operator, value: '<=', input_type: :select },
-                        { key: :season, value: main.season, input_type: :select }], path: index_path)
-        }
-        it_behaves_like :contains, true, false
-      end
-      context 'gteq season' do
-        subject {
-          ajax_actions([{ key: :season_operator, value: '>=', input_type: :select },
-                        { key: :season, value: main.season, input_type: :select }], path: index_path)
-        }
-        it_behaves_like :contains, true, true
-      end
-=begin
-      context :from_later do
-        subject {
-          ajax_action_filter(key: :season_from, value: sub.season,
-                             input_type: :select, path: index_path)
-        }
-        it_behaves_like :contains, false, true
-      end
-
-      context :to_earlier do
-        subject {
-          ajax_action_filter(key: :season_to, value: main.season,
-                             input_type: :select, path: index_path)
-        }
-        it_behaves_like :contains, true, false
-      end
-=end
-    end
-    shared_context :ajax_filter do |filter|
-      context filter.key do
-        subject {
-          ajax_action_filter(key: filter.key, value: main.send(filter.key),
-                             input_type: filter.input_type, path: index_path)
-        }
-        it_behaves_like :contains, true, false
+    shared_context :filter_season do | datatable |
+      filter = datatable.filters.flatten.find {|d| d.key == :season}
+      context 'filter_season' do
+        include_context :filter_with_operator, filter, :season_operator, '>'
+        include_context :filter_with_operator, filter, :season_operator, '<'
       end
     end
 
@@ -108,51 +116,47 @@ module AjaxFeatureHelper
     def ajax_action_filter(path:, input_type:, key:, value: nil)
       ajax_actions([key: key, value: value, input_type: input_type], path: path)
     end
+
   end
 
   ################
   module Order
     RSpec::Matchers.define :appear_before do |later_content|
       match do |earlier_content|
-        table_id = find(:css, '.dataTable')[:id]
-        table_text = find("##{table_id}").text
-        table_text.index(earlier_content.to_s) < table_text.index(later_content.to_s)
+        table_html = get_datatable(page)['outerHTML']
+        table_html.index(earlier_content.to_s) < table_html.index(later_content.to_s)
       end
-    end
-
-    shared_examples :order_main_sub do |key, identifer_key: :name|
-      it {
-        table_id = find('.dataTable')[:id]
-        dir = find("#column_#{table_id}_#{key}")['class']
-        identifers = [main, sub].sort_by { |d| d.send(key) }.map { |d| d.send(identifer_key) }
-        identifers.reverse! if dir =~ /sorting_desc/
-        expect(identifers.first).to appear_before identifers.second
-      }
     end
     ### order
-    shared_context :order do |datatable_class, excludings: []|
-      datatable_class.new.columns.select(&:orderable).map(&:name).each do |key|
-        next if excludings.include?(key.to_sym)
+    shared_context :orders do |datatable|
+      context :orders do
+        datatable.columns.select(&:orderable).map(&:name).map(&:to_sym).each do |key|
+          context key do
+            it {
+              column = datatable.columns[key]
+              expected = datatable.data.order("#{column.source} asc").map(&:name)
+              raise if expected.count < 2
 
-        include_context :ajax_order, key
+              ## asc
+              visit datatable_index_path(datatable)
+              table_id = find('.dataTable')[:id]
+              column_id = "column_#{table_id}_#{key}"
+              dir = find("#column_#{table_id}_#{key}")['aria-sort']
+              find("##{column_id}").click
+              sleep SLEEP_COUNT
+
+              expected.reverse! if  dir == 'ascending'
+              expect(expected.first).to appear_before expected.last
+
+              ## desc
+              find("##{column_id}").click
+              sleep SLEEP_COUNT
+              
+              expect(expected.last).to appear_before expected.first
+            }
+          end
+        end
       end
-    end
-
-    shared_context :ajax_order do |key, identifer_key: :name|
-      context key do
-        subject! { ajax_action_order(key, path: index_path) }
-        it_behaves_like :order_main_sub, key, identifer_key: identifer_key
-      end
-    end
-    ### ajax
-    def ajax_action_order(column_name, path:)
-      visit path
-      table_id = find('.dataTable')[:id]
-      column_id = "column_#{table_id}_#{column_name}"
-
-      find("##{column_id}").click
-      sleep SLEEP_COUNT
-      page
     end
   end
 end