+++ spec/features/scores_feature_spec.rb
@@ -1,22 +1,20 @@
 require 'rails_helper'
 
 feature ScoresController, type: :feature, feature: true do
-  let!(:main) { create(:competition, :world).scores.first }
-  let!(:sub) { create(:competition, :finlandia).scores.first }
-  let(:index_path) { scores_path }
+  #let!(:main) { create(:competition, :world).scores.first }
+  #let!(:sub) { create(:competition, :finlandia).scores.first }
+  before {
+    create(:competition, :world)
+    create(:competition, :finlandia)
+  }
+
 
   ################
   feature '#index', js: true do
-    context 'all' do
-      subject { visit index_path; page }
-      it_behaves_like :contains, true, true
-    end
-    context 'filter' do
-      include_context :filter, ScoresDatatable::Filters.new, excludings: [:season_operator]
-      include_context :filter_season
-    end
-    context 'order' do
-      include_context :order, ScoresDatatable
-    end
+    datatable = ScoresDatatable.new
+    include_context :contains_all, datatable
+    include_context :filters, datatable
+    include_context :filter_season, datatable
+    include_context :orders, datatable
   end
 end